+++ cmd/activator/activity_test.go
@@ -53,7 +53,7 @@ func TestSilentTCPConnectionIsNotActivity(t *testing.T) {
 		t.Fatal(err)
 	}
 	defer ln.Close()
-	go p.serveTCP(ln, be.Addr().String())
+	go p.serveTCP(ln, be.Addr().String(), "")
 
 	c, err := net.Dial("tcp", ln.Addr().String())
 	if err != nil {
@@ -102,7 +102,7 @@ func TestSilentTCPConnectionIsNotClosed(t *testing.T) {
 		t.Fatal(err)
 	}
 	defer ln.Close()
-	go p.serveTCP(ln, be.Addr().String())
+	go p.serveTCP(ln, be.Addr().String(), "")
 
 	c, err := net.Dial("tcp", ln.Addr().String())
 	if err != nil {